# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FeedSettingsResponse": {
        "properties": {
          "group_duplicates": {
            "title": "Group Duplicates",
            "type": "boolean"
          }
        },
        "required": [
          "group_duplicates"
        ],
        "title": "FeedSettingsResponse",
        "type": "object"
      },
      "FeedSettingsUpdate": {
        "properties": {
          "group_duplicates": {
            "title": "Group Duplicates",
            "type": "boolean"
          }
        },
        "required": [
          "group_duplicates"
        ],
        "title": "FeedSettingsUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "teleyes",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/settings/feed": {
      "get": {
        "operationId": "get_feed_settings_settings_feed_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "teleyes_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Teleyes Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feed Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_feed_settings_settings_feed_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "teleyes_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Teleyes Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Feed Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}